Move the ArrayKeyStore to the test utils namespace
