+++ tests/Util/ArrayKeyStore.php
@@ -1,12 +1,9 @@
 <?php
 
-namespace Firebase\Auth\Token;
+namespace Firebase\Auth\Token\Tests\Util;
 
 use Firebase\Auth\Token\Domain\KeyStore;
 
-/**
- * @codeCoverageIgnore
- */
 class ArrayKeyStore implements KeyStore
 {
     private $keys;
@@ -19,7 +16,7 @@ class ArrayKeyStore implements KeyStore
     public function get($keyId)
     {
         if (!array_key_exists($keyId, $this->keys)) {
-            throw new \OutOfBoundsException();
+            throw new \OutOfBoundsException(sprintf('Key with ID "%s" not found.', $keyId));
         }
 
         return $this->keys[$keyId];
